[IMP] hr_contract: Split files according to models
